* How do I create a static variable?
I'm completely replacing this answer, so I've included the new
answer as straight text, then the patch at the end.

In the new answer, I define a "static" variable, then show an
example for one function that references an out-of-scope
lexical.  The old answer didn't do that.

After that, I show that the out-of-scope lexical can be shared
with other functions (like the previous answer did), so it
isn't really "static".

I cut out most of the discussion of file-scoped variable: it's
enough to remind people that a file is a scope and that lexical
variables can't be seen outside of their scope.  If the reader
is looking for static variables, they probably aren't looking
for file-scoped variables.

I also added a pointer to information on closures, which is
where most of the valuable discussion on out-of-scope lexicals
takes place.


git-svn-id: http://svn.perl.org/modules/perlfaq/trunk@2881 50811bd7-b8ce-0310-adc1-d9db26280581