Change ${SVN} to ${VCSUPDATE}.
The problem with ${SVN} is that this works for values of SVN along
the lines of 'svn' or '/usr/local/bin/svn' or even 'cvs', but not
so well for values like 'git', since 'git update' doesn't do what
you probably want.

So, set this to an entire command line, and make the variable name
less SVN-centric. Then you can set it to 'svn update' if you'd like,
or 'git pull', or whatever you want.

Signed-off-by: brian d foy <brian.d.foy@gmail.com>